requirements.txt
matplotlib
calplot>=0.1.7.5
pandas>=1.3.0
folium
streamlit_folium
seaborn
numpy
